Create pkg testfs
