+++ internal/pkg/local/delete_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/keboola/keboola-as-code/internal/pkg/filesystem"
 	"github.com/keboola/keboola-as-code/internal/pkg/fixtures"
-	"github.com/keboola/keboola-as-code/internal/pkg/testhelper"
+	"github.com/keboola/keboola-as-code/internal/pkg/testfs"
 )
 
 func TestLocalDeleteModel(t *testing.T) {
@@ -52,7 +52,7 @@ func TestLocalDeleteModel(t *testing.T) {
 
 func TestDeleteEmptyDirectories(t *testing.T) {
 	t.Parallel()
-	fs := testhelper.NewMemoryFs()
+	fs := testfs.NewMemoryFs()
 
 	// Structure:
 	// D .hidden
