+++ internal/pkg/mapper/sharedcode/helper/helper_test.go
@@ -9,12 +9,12 @@ import (
 	"github.com/keboola/keboola-as-code/internal/pkg/log"
 	"github.com/keboola/keboola-as-code/internal/pkg/model"
 	"github.com/keboola/keboola-as-code/internal/pkg/testapi"
-	"github.com/keboola/keboola-as-code/internal/pkg/testhelper"
+	"github.com/keboola/keboola-as-code/internal/pkg/testfs"
 )
 
 func TestGetSharedCodeByPath(t *testing.T) {
 	t.Parallel()
-	fs := testhelper.NewMemoryFs()
+	fs := testfs.NewMemoryFs()
 	state := model.NewState(log.NewNopLogger(), fs, model.NewComponentsMap(testapi.NewMockedComponentsProvider()), model.SortByPath)
 	naming := model.DefaultNamingWithIds()
 	h := New(state, naming)
@@ -41,7 +41,7 @@ func TestGetSharedCodeByPath(t *testing.T) {
 
 func TestGetSharedCodeRowByPath(t *testing.T) {
 	t.Parallel()
-	fs := testhelper.NewMemoryFs()
+	fs := testfs.NewMemoryFs()
 	state := model.NewState(log.NewNopLogger(), fs, model.NewComponentsMap(testapi.NewMockedComponentsProvider()), model.SortByPath)
 	naming := model.DefaultNamingWithIds()
 	h := New(state, naming)
@@ -68,7 +68,7 @@ func TestGetSharedCodeRowByPath(t *testing.T) {
 
 func TestGetSharedCodeVariablesId(t *testing.T) {
 	t.Parallel()
-	fs := testhelper.NewMemoryFs()
+	fs := testfs.NewMemoryFs()
 	state := model.NewState(log.NewNopLogger(), fs, model.NewComponentsMap(testapi.NewMockedComponentsProvider()), model.SortByPath)
 	naming := model.DefaultNamingWithIds()
 	h := New(state, model.DefaultNamingWithIds())
